Rename ConfigUtils to IntegrationConfig; Speed up IntegrationConfig initialization
